+++ lib/src/clojure_lsp/feature/diagnostics/built_in.clj
@@ -266,7 +266,7 @@
 (defn ^:private cyclic-dependencies
   "Detects cyclic dependencies and generates diagnostics for each namespace in a cycle."
   [narrowed-db _project-db settings]
-  (let [level (get-in settings [:linters :clojure-lsp/cyclic-dependencies :level] :error)]
+  (let [level (get-in settings [:linters :clojure-lsp/cyclic-dependencies :level] :off)]
     (when-not (identical? :off level)
       (let [cycles (find-dependency-cycles (:dep-graph narrowed-db))
             exclude-namespaces (set (get-in settings [:linters :clojure-lsp/cyclic-dependencies :exclude-namespaces] #{}))