+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Change `clojure-lsp/cyclic-dependencies` custom linter default level to be `off` until corner cases are fixed.
+
 ## 2025.08.15-17.11.38
 
 - Fix `clojure-lsp/cyclic-dependencies` custom linter default level to be `error`.